requirements-dev.txt
# Development Dependencies for Dmitry

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # Parallel test execution

# Code Quality
ruff==0.1.9  # Fast Python linter and formatter
mypy==1.7.1  # Static type checker
black==23.12.1  # Code formatter (alternative to ruff format)

# Type Stubs
types-requests==2.31.0.10
types-redis==4.6.0.11

# Security
bandit==1.7.6  # Security linter
safety==2.3.5  # Dependency vulnerability scanner

# Configuration Management
pydantic[dotenv]==2.5.3  # Settings management with .env support

# Structured Logging
structlog==23.2.0  # Structured logging
python-json-logger==2.0.7  # JSON log formatter

# Observability (OpenTelemetry)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation==0.42b0

# Caching
redis==5.0.1  # Redis client
hiredis==2.3.2  # Fast Redis parser

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings[python]==0.24.0

# Development Tools
ipython==8.18.1  # Enhanced Python shell
ipdb==0.13.13  # IPython debugger
watchdog==3.0.0  # File system monitoring
pre-commit==3.6.0  # Git hooks

# Profiling
py-spy==0.3.14  # Sampling profiler
memory-profiler==0.61.0  # Memory profiler

# Mocking
responses==0.24.1  # Mock HTTP requests
freezegun==1.4.0  # Mock datetime

# Load Testing
locust==2.20.0  # Load testing tool

# Include production dependencies
-r MarkX/requirements_production.txt
